Prevent people from updating other people's chores to completed. Updated bug in profile page that made values under 0 round to 0.

+++ app/src/main/java/com/example/evan/homies/ProfileFragment.kt
@@ -128,8 +128,8 @@ class ProfileFragment : Fragment(), OnChartValueSelectedListener {
         if(numOfChores != 0) {
             entries.clear()
             chart.setTouchEnabled(true)
-            entries.add(PieEntry((completedTasks / numOfChores) * 100f, "Completed Tasks"))
-            entries.add(PieEntry(((numOfChores - completedTasks) / numOfChores) * 100f, "Incomplete Tasks"))
+            entries.add(PieEntry((completedTasks.toFloat() / numOfChores.toFloat()) * 100f, "Completed Tasks"))
+            entries.add(PieEntry(((numOfChores.toFloat() - completedTasks.toFloat()) / numOfChores.toFloat()) * 100f, "Incomplete Tasks"))
 
             chart.notifyDataSetChanged()
             chart.invalidate()
